Save points, Hero prefab modified, animation change
The game now saves every time the player triggers the objectives game object.
Changes in the scene ProjectFOLS for Hero and the position of the objectives.
Changed the Hero prefab so we know directly what is needed.
Duplicated the Melee animation so we can modify in it to put an event for sound.