Objectives modifications, Colliders added
Added Colliders at both side of the path leading to the doors to the boss in ProjectFOLS.
Objectives disappears after 5s when activating the objectives triggers and when loading the game after.
Doors are now open even after reloading the game if triggered previously.
The player is trapped in the arena when triggering the ColliderFight until he reaches the next save point.