Editor Button Added to delete save file, UI Active on Awake
An Editor Button is added in the inspector of DataPersistenceManager game object, by clicking it, it deletes the save file of the game if it exists.
UI Menu are active during Awake but inactive by default (even before runtime) -> Need to add the UI game objects in the inspector under Ui Objects in UIManager game object, and deactivate all game objects in the list in the inspector.