All scenes modified for the implementation of loading the scenes properly (still need more fixing)
For all scenes:
LevelChanger animator is put to UnscaledTime to not be affected when we pause to switch scene (x scene to main menu).
UI deactivated and put in the UI objects list in the UIManager so that no UI is obstructing anything in the game view (except MainMenu).

LoadScene script modified but still need to fix a few things for transitions between scenes.