Save and Load system updated
From the MainMenu, if there's a save file, the player will spawn at the scene he was and at the last save point he triggered. If not, a new game starts and the player will spawn at the RealHub.

From RealHub, the player will spawn from the beginning of the dungeon in ProjectFOLS when loading the scene. The save points will also reset every time the player loads the scene again from RealHub.

When the player spawns in RealHub, the game will do a save and the player stats will be maxed. The player's coins will be set to 0 and the gems and keys will keep their currencies from the last save.

Currently, removed the doors in ProjectFOLS from GameData.