Implementation of Save and Load system in NewLevel, small fixes
NewLevel has now the save and load system implemented.
Fixed loading canvas when transitioning scenes and fixed interaction UI with the chest.
WaveSpawner, CloseFightArea and ActivatePentagram objects have their data saved in the save file.
Currently when activating a pentagram, the game will be saved (need to verify something else...)
Fixed the error from the game manager trying to find input in MainMenu when there's none.
Doors in ProjectFOLS are back in the save file.
Removed InputManager object from MainMenu because not necessary.
Reset numbers of keys to 0 when loading another scene.