New arrangements and fixes
NewLevel has now a new BGM.
The Key given from Chest is now given after player collects the loot from Chest.
The Pause Menu is deactivated when returning to MainMenu.
The boss in ProjectFOLS now spawns when the door opens.
For each Pentagram area, the player is enclosed and is forced to fight the enemies.
To activate the pentagram, the player needs a key.