NewLevel UI modified
NPC Dialogue and Sign is now facing the camera and do not follow the player around.
Interactive button UI shows depending if they're talking to the NPC or not and other things related for smoother interaction.
The puzzle tip UI is showing the answer.
UI text modified and canvas as well to fit what was needed.